+++ tests/run_dcli_commands.sh
@@ -6,7 +6,6 @@ set -e
 
 export DSH_CLI_PLATFORM=nplz
 export DSH_CLI_TENANT=greenbox-dev
-export DSH_CLI_GUID=1903
 export DSH_CLI_PASSWORD_FILE=np-aws-lz-dsh.greenbox-dev.pwd
 
 export APP_UNDER_TEST=kafdrop